Edits and TODOs for chap 5 up to line 441
Please pay close attention to the edits I made. There were (and likely
still are) a lot of issues with tense and voice, and the content doesn't
always flow well from subsec to subsec, or even within subsections.
There are still issues with flow and you need to do another careful read
through for this before sending it to Peggy.